+++ oidc_provider/views.py
@@ -67,11 +67,11 @@ class AuthorizeView(View):
                 if hook_resp:
                     return hook_resp
 
-                if settings.get('OIDC_SKIP_CONSENT_ALWAYS') and not (authorize.client.client_type == 'public') \
+                if not authorize.client.require_consent and not (authorize.client.client_type == 'public') \
                 and not (authorize.params['prompt'] == 'consent'):
                     return redirect(authorize.create_response_uri())
 
-                if settings.get('OIDC_SKIP_CONSENT_ENABLE'):
+                if authorize.client.reuse_consent:
                     # Check if user previously give consent.
                     if authorize.client_has_user_consent() and not (authorize.client.client_type == 'public') \
                     and not (authorize.params['prompt'] == 'consent'):